requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gimmik==3.2.1
gitdb==4.0.12
GitPython==3.1.45
h5py==3.15.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
libxsmm==0.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpi4py==4.1.1
mpmath==1.3.0
narwhals==2.12.0
networkx==3.5
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
pooch==1.8.2
prompt_toolkit==3.0.52
protobuf==6.33.1
pyarrow==21.0.0
pydeck==0.9.1
pyfr==2.1
Pygments==2.19.2
pyobjc-core==12.1
pyopencl==2025.2.7
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytools==2025.2.5
pytz==2025.2
pyvista==0.46.4
PyYAML==6.0.3
questionary==2.1.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
rtree==1.4.1
scipy==1.16.3
scooby==0.11.0
siphash24==1.8
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.9.1
torchdiffeq==0.2.5
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vtk==9.5.2
wcwidth==0.2.14